# openapi/avedra-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AVEDRA API",
    "description": "API REST de AVEDRA — Asistencia, Convivencia y Evaluacion",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/asistencia/estados": {
      "get": {
        "tags": [
          "asistencia"
        ],
        "summary": "Estados Por Grupo Y Fecha",
        "description": "Retorna los estados de asistencia registrados para el grupo en la fecha.\ndict[estudiante_id, {estado, observacion}]",
        "operationId": "estados_por_grupo_y_fecha_api_v1_asistencia_estados_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "grupo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          },
          {
            "name": "asignacion_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Asignacion Id"
            }
          },
          {
            "name": "fecha",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Fecha"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/asistencia/masiva": {
      "post": {
        "tags": [
          "asistencia"
        ],
        "summary": "Guardar Masiva",
        "description": "Guarda la asistencia de un grupo en una fecha (en lote).",
        "operationId": "guardar_masiva_api_v1_asistencia_masiva_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardarAsistenciaMasivaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/asistencia/resumen-grupo": {
      "get": {
        "tags": [
          "asistencia"
        ],
        "summary": "Resumen Grupo",
        "description": "Retorna el resumen de asistencia de todos los estudiantes del grupo.",
        "operationId": "resumen_grupo_api_v1_asistencia_resumen_grupo_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "grupo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          },
          {
            "name": "asignacion_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Asignacion Id"
            }
          },
          {
            "name": "periodo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Periodo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumenAsistenciaResponse"
                  },
                  "title": "Response Resumen Grupo Api V1 Asistencia Resumen Grupo Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auditoria/eventos": {
      "get": {
        "tags": [
          "auditoria"
        ],
        "summary": "Listar Eventos",
        "description": "Lista eventos de sesion (login, logout, fallos) paginados.",
        "operationId": "listar_eventos_api_v1_auditoria_eventos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "usuario_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Usuario Id"
            }
          },
          {
            "name": "tipo_evento",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipo Evento"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 25,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_EventoSesionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auditoria/cambios": {
      "get": {
        "tags": [
          "auditoria"
        ],
        "summary": "Listar Cambios",
        "description": "Lista cambios del audit_log paginados.",
        "operationId": "listar_cambios_api_v1_auditoria_cambios_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "usuario_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Usuario Id"
            }
          },
          {
            "name": "tabla",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tabla"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 25,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_RegistroCambioResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auditoria/integridad": {
      "get": {
        "tags": [
          "auditoria"
        ],
        "summary": "Verificar Integridad",
        "description": "Verifica el encadenamiento por hash de las tablas de auditoria.",
        "operationId": "verificar_integridad_api_v1_auditoria_integridad_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "completa",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Completa"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegridadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EstadoAsistenciaItem": {
        "properties": {
          "estudiante_id": {
            "type": "integer",
            "title": "Estudiante Id"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          },
          "observacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observacion"
          }
        },
        "type": "object",
        "required": [
          "estudiante_id",
          "estado"
        ],
        "title": "EstadoAsistenciaItem",
        "description": "Item de asistencia para guardar en lote."
      },
...
      "EventoSesionResponse": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "usuario": {
            "type": "string",
            "title": "Usuario"
          },
          "usuario_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usuario Id"
          },
          "tipo_evento": {
            "type": "string",
            "title": "Tipo Evento"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "fecha_hora": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha Hora"
          },
          "detalles": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detalles"
          },
          "severidad": {
            "type": "string",
            "title": "Severidad",
            "default": "INFO"
          },
          "institucion_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institucion Id"
          }
        },
        "type": "object",
        "required": [
          "usuario",
          "tipo_evento",
          "fecha_hora"
        ],
        "title": "EventoSesionResponse"
      },
...
      "GuardarAsistenciaMasivaRequest": {
        "properties": {
          "grupo_id": {
            "type": "integer",
            "title": "Grupo Id"
          },
          "asignacion_id": {
            "type": "integer",
            "title": "Asignacion Id"
          },
          "periodo_id": {
            "type": "integer",
            "title": "Periodo Id"
          },
          "fecha": {
            "type": "string",
            "format": "date",
            "title": "Fecha"
          },
          "lista": {
            "items": {
              "$ref": "#/components/schemas/EstadoAsistenciaItem"
            },
            "type": "array",
            "title": "Lista"
          }
        },
        "type": "object",
        "required": [
          "grupo_id",
          "asignacion_id",
          "periodo_id",
          "fecha",
          "lista"
        ],
        "title": "GuardarAsistenciaMasivaRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IntegridadResponse": {
        "properties": {
          "eventos_ok": {
            "type": "boolean",
            "title": "Eventos Ok"
          },
          "cambios_ok": {
            "type": "boolean",
            "title": "Cambios Ok"
          },
          "evento_roto_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evento Roto Id"
          },
          "cambio_roto_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cambio Roto Id"
          },
          "alcance": {
            "type": "string",
            "title": "Alcance",
            "default": "incremental"
          }
        },
        "type": "object",
        "required": [
          "eventos_ok",
          "cambios_ok"
        ],
        "title": "IntegridadResponse"
      },
...
      "PaginatedResponse_EventoSesionResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventoSesionResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "per_page"
        ],
        "title": "PaginatedResponse[EventoSesionResponse]"
      },
      "PaginatedResponse_RegistroCambioResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RegistroCambioResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "per_page"
        ],
        "title": "PaginatedResponse[RegistroCambioResponse]"
      },
...
      "RegistroCambioResponse": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "usuario_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usuario Id"
          },
          "usuario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usuario"
          },
          "accion": {
            "type": "string",
            "title": "Accion"
          },
          "tabla": {
            "type": "string",
            "title": "Tabla"
          },
          "registro_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registro Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "institucion_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institucion Id"
          }
        },
        "type": "object",
        "required": [
          "accion",
          "tabla",
          "timestamp"
        ],
        "title": "RegistroCambioResponse"
      },
...
      "ResumenAsistenciaResponse": {
        "properties": {
          "estudiante_id": {
            "type": "integer",
            "title": "Estudiante Id"
          },
          "total_clases": {
            "type": "integer",
            "title": "Total Clases",
            "default": 0
          },
          "presentes": {
            "type": "integer",
            "title": "Presentes",
            "default": 0
          },
          "faltas_justificadas": {
            "type": "integer",
            "title": "Faltas Justificadas",
            "default": 0
          },
          "faltas_injustificadas": {
            "type": "integer",
            "title": "Faltas Injustificadas",
            "default": 0
          },
          "retrasos": {
            "type": "integer",
            "title": "Retrasos",
            "default": 0
          },
          "excusas": {
            "type": "integer",
            "title": "Excusas",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "estudiante_id"
        ],
        "title": "ResumenAsistenciaResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}